refactor code for quick start
